backend/requirements.txt
flask
flask-cors
PyPDF2
transformers
supabase
python-dotenv
gunicorn